filter_grep: use internal regex engine instead of regex.h (#502)
the old implementaion of filter_grep uses the system regex.h
engine instead of libonigmo which is the default solution.

This patch deprecate the usage of regex.h.

note: this plugin should behave as Fluentd grep plugin, it needs
to be fixed to work with rules in 'pipe' mode instead of a simple
ACL.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>