in_tail: check if file is already being monitored (#486 #508)
there are certain scenarios when log rotation is triggered by a third
party application, in_tail perform a direct new file registration
just before the scan routine. This is right, which is wrong is not to
check the file is not registered before. The bug generate the following
unexpected behaviors:

 - after long periods and high number of rotations, memory goes up.
 - file descriptors may not be closed properly for duplicates

The following patch adds an extra validation when a new file is going
to be added to the queue of monitored files (do not add duplicates).

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>