Plugin: record_modifier: Suffix_Type modification for keys
This patch adds an option to record_modifier plugin that modifies
record key names by suffixing a type identifier.

It is activated with the boolean option Suffix_Type (default False).

For example if processing a JSON log message like:

{"port":8000,"app":"myapp","ssl":true}

it would generate key names: port_i, app_s and ssl_b

This is useful primarily for JSON structured log aggregation output
to Elasticsearch. Once the data type for a key is set for an index it
can't be changed, which can lead to conflicts if different data types
are used for the same key in a large application or across different
applications.

With this patch {"port":"8000"} and {"port":8000} result in different
key names (port_s and port_i) respectively, which prevents data type
conflicts.

Signed-off-by: Rory Douglas <rory1douglas@gmail.com>