Additional json_lines Format for HTTP Output
This is an attempt to provide a simple option that allows JSON
logs to be sent to services that require each log entry be on
a separate line.  An example is Loggly.  Currently testing this.

Avoid having the first line be a newline.

This is more consistent with #577 as well.

Signed-off-by: Brandon Cole <brandon.w.cole@gmail.com>