out_splunk: Allow user to send raw message by setting Splunk_Send_Raw On
This patch adds a new toggle for the Splunk HTTP Event Collector output
plugin to allow users to specify index, sourcetype and other top level
fields when sending data to splunk.  The current behaviour sends data to
the main index by default.

When enabling the feature, the user must take care to put all log
details in the event field, and only specify fields known to splunk in
the top level event.

Example:

    Splunk_Send_Raw Off
    {"time": .., "event": {"k1": "foo", "k2": "bar", "index": "applogs"}}

    Splunk_Send_Raw On
    {"time": .., "k1": "foo", "k2": "bar", "index": "applogs"}

For up to date information about the valid keys in the top level object,
refer to the splunk documentation:

http://docs.splunk.com/Documentation/Splunk/latest/Data/AboutHEC

The patch has been tested using Splunk 6.6.1

Signed-off-by: Carl Henrik Lunde <chlunde@ifi.uio.no>