flb_pack: fix crash with stricter JSON parsing
JSMN is not a strict JSON parser, for example

    1 {"foo": 1} false

will generate a stream of JSON tokens.  This will confuse other layers
of fluent-bit which might lead to crashes.  In particular, the
kubernetes filter will often crash with such data.

This change validates that there's only a single JSON object on the
line, and that this object spans the whole line except for whitespace
before or after the JSON object.

I've ran fluent-bit with americal fuzzy lop (AFL) for several hours
without any crashes.

Fixes #611, #584.

Signed-off-by: Carl Henrik Lunde <chlunde@ifi.uio.no>