out_es: new 'replace_dots' option (#708)
On Elasticsearch 2.0-2.3, key names with dots were not allowed for
hence was required to replace every dot with an underscore. This
requirement is not longer true on newer versions of Elasticsearch.

The following patch introduce a new option called 'replace_dot' which
is disabled by default. So now the plugin will only replace the dots
if the flag is enabled for use cases where old versions of Elasticsearch
are in use.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>