filter_kubernetes: re-enable 'merge_log_key' and add new 'merge_log_trim' (#715)
This patch re-enable the old disabled feature called 'merge_json_key' with a
new name as 'merge_log_key', when enabled all parsed structured content from
the 'log' field is added as part of a map with the given name.

In addition a new option called 'merge_log_trim' is added to remove any trailing
breakline (\n) or carrier return (\r) characters at the end of a structured
log value, this flag is enabled by default.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>