in_tail: fix filename lookup in OSX (#757)
When solving the file name associated to a file descriptor, there
are different ways to get it depending on the operating system we
are running. On Linux this is done through ProcFS and the returned
value respect original paths which do not expand if the root path
is a symbolic link; in the other hand in OSX a fcntl() with F_GETPATH
is required and if the opened file is under a symbolic link it will
expand to the real path generating some issues when detecting rotated
files.

This patch adds an extra 'real_name' field to the file structure to
keep consistency and do a double check for paths fixing the issues
found on rotation and file offsets. This fix is applicable only for
OSX.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>