filter_kubernetes: fix timestamp lookup on parsed record through annotation (#763)
When a Pod suggest a parser through fluentbit.io/parser annotation and the record
is parsed, Fluent Bit was ommitting the timestamp discovered by the parser and
only setting up the timestamp set by the original JSON map generated by Docker.

This patch fix the wrong behavior, if a parser succeed and a new timestamp is found,
use that one, otherwise fallback to the timestamp set by Docker.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>