Add ability to preload Kubernetes meta data from directory
A new configuration parameter (Kube_Meta_Preload_Cache_Dir) points
to a directory that expects JSON files named as:
`namespace-pod.meta`. If present, this is used instead of fetching
from the API. Note: we could refactor the tests to use this
instead of the in-built HTTP server.

```
[FILTER]
    Name                kubernetes
    Match               kube.*
    Merge_JSON_Log      On
    K8S-Logging.Parser  On
    K8S-Logging.Exclude On
    Kube_Meta_Preload_Cache_Dir meta
```

The Kubernetes meta tag-regex is improved to allow arbitrary locations
(rather than having var/log/containers hard-coded). This allows
both this feature, but also for Kubernetes environments that store
logs in a different directory to function.

In tail_file, the inital tag is made from 'basename(file)' rather
than full path. This simplifies the parsing which would otherwise
have to look for an arbitrary number of leading dot-delimited fields
(since the escape mechanism changes / to .).

Signed-off-by: Don Bowman <don@agilicus.com>