Build: Adding upstart init system support
This patch provides an upstart job file to be passed to the build process;
The job's config mimics the behaviour of the systemd service.
CMake config modified to handle the above.
Also, there was some duplication of '/' in the resulting package's
paths, that is fixed as well.

Build tested with modified 'fluent-bit-packaging' project;
Resulting package tested on Ubuntu 14.04.

Signed-off-by: Zsolt Kulcsar <zskulcsar@gmail.com>