Add optimisation / partial-work-around for non-JSON in 'Merge_Log' case
In the case we are using Kubernetes and Merge_Log = On, if a message
is logged in which the first character is a valid token, but it is not
itself a valid JSON object, the parser will lock up. E.g. if a message
of '----' is logged, the first char will pass the tokenisation test
in jsmn, and tokens!=NULL since the outer 'docker' format is JSON
and mid-parse. This in turn means the flb_pack_json will never finish
and the input will lock up. Add a simple check for '[' or '{' which
are the only valid start-of-object characters. Note: this is only
a partial work-around, since a message could be { ... w/o closing }.
More work is needed, we need a separate jsmn parser for the inner.

Signed-off-by: Don Bowman <don@agilicus.com>