pack: flb_pack_json() adds fourth argument to store initial JSON type
There some scenarios where the caller (core / plugin) needs to validate
that the packaged json->msgpack data belongs to a specific JSON type. As
of now the only solution was to unpack the recently packed data or do
some custom validations in the JSON input string.

The following patch alter the function prototype of flb_pack_json() adding
a fourth argument called 'root_type':

  int flb_pack_json(char *js, size_t len, char **buffer, size_t *size,
                    int *root_type)

if the json parsed succeeded, it will store in root_type the specific
JSON data type found at the begining, it can take one of the following
values:

  FLB_PACK_JSON_UNDEFINED
  FLB_PACK_JSON_OBJECT
  FLB_PACK_JSON_ARRAY
  FLB_PACK_JSON_STRING
  FLB_PACK_JSON_PRIMITIVE

note: it's up to the caller do further validations and checks about the
expected data returned. If the returned root_type is not the expected,
make sure to release the msgpack buffer created.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>