Switch to distroless as base image for container, add .dockerignore
The copy-in of the code is moved to after the setup of the build
container to maximise cache (so no rebuild of the build packages
if the source changes).

Add a .dockerignore of .git / build / Dockerfile to maximise
cache.

Switch to gcr.io/distroless/cc as a base container for the run-time,
reducing the possibility of an 'escape' in the code.

Signed-off-by: Don Bowman <don@agilicus.com>