Switch travis UT to 'xenial', add matrix of clang/gcc for sanitizers
Switch to 'xenial' from 'trusty' for unit testing since the older
binutils did not work well with the sanitizer tests for gcc.
There was also a bug in libc gmtime() which was causing faulse failures.

Add a matrix so that we run, for both gcc and clang:
- JEMALLOC on/off
- ASAN on/off
- USAN on/off
- MSAN on/off

TSAN is disabled as there are significant errors/false positives
in the code where sleep() is used to synchronise.

Infrastructure is put in place to build a test docker image
with a bash test suite prepatory to committing a 'corpus' of
input test files to compare output for system test of the
docker image.

Signed-off-by: Don Bowman <don@agilicus.com>