tests: internal: de-duplicate code in input runtime tests and add a 'WITHOUT' test option (#882)
The in_* tests were all cut and paste of each other with 1 little
change each. This makes a 'core' test with wrapper functions
that call it. For simplicity of execution this is presented to cmake
as separate files (so that execution is separate).

Since not all tests are 'healthy' a WITHOUT_ option is addd to
cmake for the tests, allowing one to use 'ctest' to run all tests
rather than maintaining a script to know which to skip.

The override of test compile optimisation is also removed as it
interferes with the sanitizer unit testing.

Signed-off-by: Don Bowman <don@agilicus.com>