in_tail: fix high CPU usage on busy notification channels (errno related)
There are special cases where the notification channels get busy and the
internal signal routines is relying on a previous 'errno' status to
consume the notification byte. The problem is that errno could be set
by any other component in the stack before hit that call.

Standard libraries are not forced to set errno to zero before it work so
it cannot be trusted to perform conditional operations, there is a good
reading here about that:

https://wiki.sei.cmu.edu/confluence/pages/viewpage.action?pageId=87152351

This patch makes to consume the notification byte at least once, so after
the call the validation of errno v/s EAGAIN is handled properly.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>