test: filter_kubernetes: send systemd test message after start
The filter_kubernetes test of systemd journal was sometimes
failing. The systemd input plug would do a 'tail', which, depending
on timing, would be before or after the test message we primed with.

Send the test message after the flb_start() to avoid this.

Secondly, in some CI environment, systemd is not properly running.
Do an assessment of the journal health and skip the journal
test if it is not running.

Signed-off-by: Don Bowman <don@agilicus.com>