flb_pack: improve the buffer allocation strategy
When serialization fails, we try to add a 128-byte block to the
buffer and retry again. This strategy seemingly works fine.

Except when we have an input with high flow volume, and suddenly
the consent growth factor (128 bytes) becomes too small. And we
end up wasting time doing a ridiculous amount of recalculations:

| EVENTS | INITIAL [bytes] | ACTUAL [bytes] | RECALC [calls] |
| ------ | --------------- | -------------- | -------------- |
|      1 |             36  |            47  |             1  |
|     10 |            349  |           470  |             1  |
|    100 |           3483  |          4700  |            10  |
|   1000 |          34803  |         47000  |            96  |
|  10000 |         348003  |        470000  |           954  |

This patch fixes the issue by using an exponential growth factor,
instead of a constant one.

Also I adjusted the initial buffer size to increase the chance
of getting a complete result in the first time.

Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>