task: copy tag on task creation (#967)
With the introduction of the new storage layer, tasks had a reference
to the Tag coming from chunkio buffer. This affected routing since tags
were not NULL terminated and also a possible future issue where a task
could  reference a tag coming from a memory-map which is still not loaded
can lead to a corruption.

For safety we get back to the original approach where Tasks have a copy
of the Tag they are referring to.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>