router: fix invalid routing due to invalid memory access (#967)
Recently the router was adapted to use non-NULL terminated buffers
coming from a Task that was referencing a ChunkIO memory map. This
generated some issues since the logic of the matching routine was
not ported properly plus a future issue with referenced Tags could
lead to a corruption.

This patch gets back to the original mechanism where now a Tag
contains a NULL byte at the end.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>