build: add custom build configuration to compile onigmo on Windows
I created custom build configuration for onigmo, following the install
guide in the official manual. With this patch, we are now able to use
regular expressions on Windows (both on x64 and on x86).

Note: the diff might seem cluttered, but running git-diff with `-w`
option should reveal the core difference fairly well.

Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>