filter: compose safety Tag and fix write offset (#993)
Filter was trusting that input plugins was ingesting a NULL terminated
value, but if the data comes from Forward this will not be the case and
when calling the Routing will not match any rule.

This patch add a safety tag composer plus fix a problem when setting up
the write position when data has been removed.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>