build: make LuaJIT compilable on Windows
With this change LuaJIT can be compiled and linked on MSVC.

It turned out that the Windows support was fairly streight forward
since LuaJIT already provided a working build script (msvcbuild.bat).
So what I basically did was to tell cmake to invoke this bat file,
and it just worked fine.

Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>