in_stdin: prevent double free / use after free of msgpack buffer
Currently the msgpack sbuffer is destroyed instead of cleared after
creating the msgpack objects. This is changed to clear the buffer
instead, so re-using it does not refer to deallocated memory.

Alternatively the buffer could be re-initialized after destroying it.
But since the memory, which would be used, is effectively bound to the
stdin buffer size, clearing gives a little better performance at the
expense of some small additional memory usage.

Signed-off-by: Dominik del Bondio <dominik.del.bondio@bitextender.com>