config: fix SIGSEGV errors on shutdown of input plugins (#1081)
The point is that we do not need to call both `mk_event_del()` and
`mk_event_timeout_destroy()` on the same event object. In fact,
calling both  can lead to SIGSEGV in some backend (e.g. libevent).

This patch fixies the issue by moving the call of mk_event_del() to
the "else" block.

Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>