parser: do not raise errors instantly on '%z'/'%Z' on Windows
Right now, the follwing results in a startup error on Windows:

    PS> .\bin\fluent-bit -R ..\conf\parsers.conf ...

This is just because flb_parser is too strict on tm_gmtoff support,
and would abort if SOME of the parser definitions include '%Z'/'%z'
in their Time_Format.

We relax the restriction here, and allow to use a timezone format
specifier if our portable strptime implementation (which can parse
'%Z' and '%z') is in use.

Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>