in_tail: implement flb_tail_file_name() for Windows
This is part of the ongoing effort to port in_tail to Windows.

This patch teaches in_tail how to get the file name from an open file
descriptor on Windows. Essentially, we needed to convert the file
descriptor to HANDLE and ask the metadata about it though Win32 API.

Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>