in_tail: use Win32 glob emulation on Windows
It really boils down to switch glob implementation for each platform.

    tail_scan.c
    ├── tail_scan_win32.c   ... Windows
    └── tail_scan_glob.c    ... Linux, BSD and OSX

To archive this, I moved the content of tail_scan.c to tail_scan_glob.c,
and set up tail_scan.c as a interface file to load an appropriate file.

With this patch in_tail can be compiled and linked on Windows.

Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>