tail: fix SIGSEGV due to the incompatible type length (int/intptr_t)
This is essentially the same fix as 13166dfb.

Since these pipe fds are actually sockets on Windows, we need to
use a 64-bit type (intptr_t) instead of 32-bit one (int) to hold
the descriptor properly.

Otherwise memory can silently get corrupt due to the type-length
mismatch, resulting in SIGSEGV.

Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>