in_tail: avoid double freeing the same timeout event object
If the input plugin supports cb_pause() interface, we ended up freeing
the same timeout object twice while shutdown, because of the function
flb_input_pause_all() destroying the timeout object in order to halt it.

This issue has been causing SIGSEGV on libevent backend (and probably
on select backend).

Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>