in_tail: add tail_scan_win32.c for scanning paths on Windows
This is a first cut at supporting wildcard pathspecs on Windows.

Since Windows has no support for glob(3), we need to emulate the
feature using Win32 API. Here I implemented it using the Win32's
FindFirstFile/FindNextFile functions, and this enables you to do:

    PS > fluent-bit -i tail -p 'path=log\*.json' -o stdout

The biggest limitation right now is that we can't use a wildcard
in the directory path (like 'C:\*\test.log'). This is because
a) FindFirstFile() does not support it and b) there seems to be no
other API that is usable for this purpose.

We'll revisit this issue in future, but the current implementation
should be sufficient for now.

Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>