in_tail: emulate inodes on Windows
in_tail uses inodes to identify files in a SQLite DB. The problem
was that Windows did not support inode and always set st_ino to 0
on stat(3) calls. Because of such, the offset storing feature was
not working properly on Windows.

This patch adds a simple emulation of inodes for Windows, and
enables Windows users to retain the read offset between restarts.

Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>