out_es: Fix elasticsearch_error_check, add trace_error output
This change does two things. In elasticsearch_error_check()
it was incorrectly returning 'was an error' all the time
(e.g. whenever the response msgpack *could* be unpacked).

Add a 'trace_error' flag to elasticsearch output to get
a raw dump of the json messages that cause a failure, in order
to debug why the failure occurred.

Signed-off-by: Don Bowman <don@agilicus.com>