out_stdout: call fflush(stdout) at the end of cb_stdout_flush()
Some terminal programs buffer stdout in memory before printing, so
emitted records may not appear on the screen promptly.

This is very confusing especially when debugging ("Is there something
wrong with my fix? Or is it just due to buffering of I/O?).

Let's flush stdout in each cb_flush call to avoid confusion.

Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>