out_es: fix buffer overrun on Trace_Output=On
While testing out_es on Windows, I noticed that Fluent Bit sometimes
outputs a garbled noise that apparently is a memory dump.

It turned out that it was just because elasticsearch_format() tried
to do printf("%s", buf) where buf is actually not null-terminated.
This patch is a simple fix for it.

Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>