ci: clean up and simplify AppVeyor's build script
This is a preparation for introducing Windows installer generation.

The previous version tries to be smart by constructing the build
option dynamically which results in messy option handling code.

This patch expands the option handling code, removes bogus flags
from there, and adds a missing flag that was actually required.

Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>