build: define FLB_SYSTEM_WINDOWS to detect Windows build target
This introduces a new macro that identifies Windows target builds
portably (it should work for both MSVC and MINGW).

Cygwin is not included here since it actually does not have many
things common with the native Windows system (it does not expose
Win32 API for starters).

From now on, we can use this macro to detect Windows platform in
the source tree.

Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>