lib: rbtree: Clone source files
This just copies files from lib/monkey/deps/rbtree.

The reason for doing this is that there are many circumstances
where we don't compile monkey (see FLB_HTTP_SERVER in CMakeList.txt),
hence we cannot always count on the existence of monkey/deps.

In those cases, we need to compile and link rbtree by ourselves, so
this commit comes in.

Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>