core: allow use of relative paths in configuration files on Windows
Fluent Bit has a feature that allows users to use a relative path
in the main configuration file. For example, if you have the
following files in the same directory:

    PS > ls C:\Program Files\td-agent\conf\
    parsers.conf fluent-bit.conf

Then you can just set "parsers.conf" to Parser_File as below:

    [SERVICE]
      Flush       5
      Daemon      Off
      Parser_File parsers.conf

Previously this feature did not work properly on Windows, due to
its path schema being incompatible with Unix. This fixes it.

Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>