out_forward: make Shared_Key usable without TLS (and vice versa)
Right now, Fluent Bit forces us to set up a common secret (Shared_Key)
if TLS/SSL encryption is in use. The opposite is also true and we can't
use a common secret on a plain-text connection.

These two features are really not dependent on each other, and should
be usable independently.

With this, we can just use Shared_Key without enabling TLS.

      [Output]
      Name forward
      Host 127.0.0.1
      Self_Hostname example.com
      Shared_Key some-secret-key

Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>