filter_kubernetes: Add a dedot option for filter kubernetes
Signed-off-by: Kyâne Pichou <kyane@kyane.fr>

When log messages contains a "." Elasticsearch may refused them.
The out_es plugin already allow to convert "." to "_" in order
to fix this issue.
Because the Kubernetes filter plugin allow to collect metadata
from Kubernetes API, and because it is a good practice to have
annotations names with namespace (write as FQDN) we should allow
to convert those "." directly from the filter_kubernetes plugin.