in_storage_backlog: upon start, put chunks 'down'
When the plugin starts, it get the storage path content (chunks) that exists in
the filesystem. Before the this patch it always kept in memory the chunks loaded
and file descriptors open which was generating some resources issues if the
storage path have many files to consume.

With this patch, chunks are putted down and only are back in memory when they
are enqueued for the engine.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>