filter_kubernetes: always add meta extracted from the record itself
Some meta are possibly extracted from the record itself (for example the
namespace name, the pod name and the container name can be extracted from the
tag). If the full meta could not be fetched from the API server, no meta at all
were added to the record. This commit changes that behaviour to add at least
the meta extracted from the record itself.

This commit also fixes two memory leaks in the code which has been modified.

Signed-off-by: Yann Soubeyrand <yann.soubeyrand@gmx.fr>