io: tls: support SNI routing for outputs
Addresses #1312.

When sending logs to TLS-enabled outputs, it may be ideal to route
based on a virtual hostname of the origin server. This is possible
with SNI. This commit adds a new setting `tls.vhost` for the secure
forwarding mode. By default, the virtual hostname will be set to
the origin hostname during the TLS handshake.

Signed-off-by: Andrey Kaipov <andreykaipov@users.noreply.github.com>