in_forward: fix heap overflow via negative length (FLU-01-003)
note: based on Fluentd ecosystem security audit report / CNCF

"In fluent-bit’s in_forward plugin, it was possible to spot an exploitable
remote heap buffer overflow vulnerability. This happens due to incorrect
handling of faulty msgpack payloads.

In the call to memmove, it is assumed that conn->buf_len is always larger
than all_used. Therefore, both values are subtracted to calculate the necessary
length for the memmove operation in order to continue with the next msgpack
packet. However, the payload below shows that this assumption can be broken."

In this patch we ensure that buffer length is always greater than consumed
bytes to avoid overflow.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>