in_forward: fix DoS via incomplete EXT format (FLU-01-004)
note: reported by Security Audit / CNCF

The problem is an invalid memory access due to the proper missing check
of return value when calling msgpack_unpacker_next_with_size() function.

The following patch adds further checks for parsing and memory errors,
it prints out the error message, cleanup the buffers and close the
connection if some of the conditions mentioned are matched.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>