in_tail: allow firstline to end with an int
The fix for #1368, applied in 826c84f forbids the firstline to end with an integer. This is required if the continuation line does not match any parser.
In cases where the line is not appended to the parent, we are save to process the line, even if the first line ends with an integer.

Signed-off-by: Thomas Berger <loki@lokis-chaos.de>