parser: decoder: json: do not unescape and skip empty spaces (#1278)
Due to recent improvements in the JSON -> Msgpack encoder, it's not
longer necessary to run a new 'unescape' routine since JSON strings
are now properly converted.

This patch makes to workaround cases where the stringified JSON map
have leading empty spaces at the beginning, e.g:

  {"log": "   {\"k\": 123}"}

So when decoding 'log' value as JSON it will be converted properly
to a map.

note: issue described on #1278.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>