filter_kubernetes: new 'merge_parser' option
When Merge_Log is enabled, the filter will try to parser the
content value of 'log' key, either using a custom parser
defined through a Pod Annotation or using the default JSON
parser.

This patch implements a new option to add a custom parser
through the configuration 'merge_parser' that can be used for
testing purposes or to set something to try before JSON, the order
of processing will be:

  - Annotation Parser (if set)
  - Merge_Parser (if set)
  - JSON

Note that Merge_Parser must be a parser already set in the
parsers configuration file.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>