filter_kubernetes: do not unescape and fix handling of a merged map (#1278)
The filter used to unescape a log value before trying to pack it as a
JSON->Msgpack map, this is not necessary and avoid some issues when merging
the log.

In addition it add a fix where an original log map generated corruption if
the option merge_log_key was set.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>