input: new 'routable' option for input plugins
Every input plugin now can be set as 'routable'. This means that the
records generaed by the input plugin could be routed to output plugins
or not.

By default all of them are routable (as previous versions), but there
are certain cases where we don't want them to be routed: consider the
case where the Stream Processor is creating new streams of data but
we don't want to keep the original records.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>