plugin: unify plugins loader and support dynamic C plugins
The following patch introduce a new formal 'plugin' interface
to handle the registration of dynamic plugins. It also inherits the
ability from the proxy interface to load plugins defined in the
configuration file.

This interface is backward compatible with the Golang proxy, and
upon plugin load request it will detect based on the plugin name
if it will be handled as a normal Fluent Bit plugin or through
the proxy for Golang shared lib.

No changes are required for plugins code, static and dynamic C
plugins have the same registration and API usage.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>