out_tcp: new TCP output plugin with optional TLS support
The following patch introduce a new TCP output plugin that can be
used with TLS enabled if required.

The payload of this plugin is through the 'format' property, it support
the following modes:

 - format = msgpack (default) | json | json_lines | json_stream

in addition if the payload is formatted in JSON, the record timestamp
can be composed with the 'json_date_format' property:

 - json_date_format = double | epoch | iso8601

the timestamp key name in JSON mode can be set using the 'json_date_key'
property.

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>