lib: mpack: add new 'mpack' MessagePack (lite) library
Fluent Bit uses MessagePack data serialization format for it internal
representation of records. In order to pack and unpack data we use
'msgpack-c' library. Despite the library is really good and have a good
performance, there are certain scenarios where we need something more
lightweight.

It's common that in the data pipeline there are many stages where we need
to count the actual number of records in a chunk, that number can be altered
by filters and also internal metrics routines needs to perform the same
calculation. I could not find a way to perform 'zero-copy' unpacking of
data just to count the records, a zone must be always allocated and data
copied everytime, which is unnecessary for our use case.

'mpack' is a lightweight MessagePack library that offers an API that for
certain operations don't need a buffer and offers a better performance,
some local tests shows a gain of 20% in speed.

This patch integrates 'mpack' and further patches will wrap functions
on top of it, specifically on flb_mp.c file.

refs:

- https://github.com/ludocode/mpack

Signed-off-by: Eduardo Silva <eduardo@treasure-data.com>