out_file: add support for arbitrary templating
This adds a new format option "template" that allows us to use
Python-like format specifiers. Here is a configuration example:

    [OUTPUT]
      Name     file
      Format   template
      Template {time} [{level}] {message}

This produce an output like the following:

    1564455346.000257 [INFO] Service is starting up.
    1564455347.000254 [INFO] Waiting when the pool is empty.

This helps us getting closer to Fluentd's formatting ability.
Notably, this allows us to do single_value formatting that is
currently missing in Fluent Bit.

Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>