in_collectd: Implement a new input plugin for collectd
This allows fluent-bit to receive UDP packets from collectd's Network
plugin. You can use it as:

    $ fluent-bit -i collectd -o stdout

... then it listens to `0.0.0.0:25826` and consumes incoming packets.

This implements the basic support for collectd's binary protocol and
also contains a compatible parser for types.db (5), so it should
cover the most real-life use cases.

The remaining features to be implemented are advanced features like
AES encryption and HMAC signatures (introduced in v4.7), which we
will address in the future patches.

Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>