in_syslog: add 'Unix_Perm' option
It solves a common pain point when using the `in_syslog` plugin. Here
is an example usage:

    [INPUT]
      Name      syslog
      Tag       syslog_data
      Parser    syslog-rfc3164
      Path      /tmp/fluent-bit.sock
      Mode      unix_udp
      Unix_Perm 0777

Without this, users are required to keep track of restarts of fluent-bit
and manually set file permission on the UNIX socket it creates. Let's
just make it work seamlessly without such manual interventions.

Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>