flb: ssl: add a new internal interface for TLS connections
`flb_ssl` is a simple interface for creating TLS connections, both
for servers and clients. The goal is to wrap mbedTLS for use inside
Fluent Bit, following the design principle of libtls.

Right now, only a set of functions that are sufficient to run a
minimum TLS server are implemented. I'm planning to expand this
library and slowly replace the existing TLS-related code base.

Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>