in_tail: prevent file duplication on macOS
Call flb_tail_file_name() for a normalized path and then perform a
duplication chack using the returned path. We normally don't need to
do this, but we _do_ need in order to handle symlinks properly on OSX.

    $ ls -l /tmp/nginx.log
    20 Sep 26 18:15 /tmp/nginx.log -> /private/tmp/nginx.log

This should resolve the bug that in_tail registers the same file again
and again.

Signed-off-by: Fujimoto Seiji <fujimoto@clear-code.com>